fix qMax calls on platforms where qreal is not double
svn path=/trunk/KDE/kdebase/workspace/libs/plasma/; revision=821243

+++ applet.cpp
@@ -599,8 +599,8 @@ void Applet::setBackgroundHints(const BackgroundHints hints)
         d->background->getMargins(left, top, right, bottom);
         //Setting a minimum size of 0,0 would result in the panel to be only
         //on the first virtual desktop
-        setMinimumSize(qMax(minimumSize().width() - left - right, 1.0),
-                       qMax(minimumSize().height() - top - bottom, 1.0));
+        setMinimumSize(qMax(minimumSize().width() - left - right, qreal(1.0)),
+                       qMax(minimumSize().height() - top - bottom, qreal(1.0)));
 
         delete d->background;
         d->background = 0;
